ventura_backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-bootstrap5
firebase-admin
nltk
Pillow
